LayoutTests/http/tests/media/video-redirect.html

<html>
<head>
</head>
<body onload="startTest()">
<video id="video" src="http://127.0.0.1:8000/media/resources/redirect-to-counting-captioned.mov"
/>
<script src=../../media-resources/video-test.js></script>
<script>
    function startTest() {
        waitForEventAndFail('error');
        waitForEventAndEnd('canplay');
        video.load();
    }
</script>
Tests that the media player will load the url redirected from reftest.mov.<br/>
</body>
</html>
