Safari-609.1.11.1/LayoutTests/http/tests/media/video-redirect-expected.txt


EVENT(canplay)
END OF TEST
